Trade like an insider
Among many participants in the US stock market, there is a unique group of investors who has knowledge and information that other investors cannot get access to. These are the company "insiders". In this study, we attempt to take a closer look into insiders buying and selling activities. We want to answer the following question: does the possession of non-public information give insiders an edge in trading?

Each section of the website is designed to tell one part of the story.  There are demo videos in each section if users need some guidelines.  

This visualization is created by Thaddeus Segura, Ying Hua, Tina Agarwal and Shirley Deng